+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Actions/SavePreviousInput.cs
@@ -7,9 +7,9 @@ using Microsoft.Bot.Builder.Solutions.Contextual.Models.Strategy;
 
 namespace Microsoft.Bot.Builder.Solutions.Contextual.Actions
 {
-    public class SavePreviousQuestion : ISkillContextualActions
+    public class SavePreviousInput : ISkillContextualAction
     {
-        public SavePreviousQuestion(
+        public SavePreviousInput(
             ConversationState convState,
             UserState userState,
             UserContextManager userContextResolver,
@@ -55,8 +55,8 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual.Actions
 
         private async Task InitPreviousQuestion(ITurnContext turnContext)
         {
-            var questionAccessor = UserState.CreateProperty<List<PreviousQuestion>>(string.Format("{0}Questions", SkillName));
-            var questions = await questionAccessor.GetAsync(turnContext, () => new List<PreviousQuestion>());
+            var questionAccessor = UserState.CreateProperty<List<PreviousInput>>(string.Format("{0}Questions", SkillName));
+            var questions = await questionAccessor.GetAsync(turnContext, () => new List<PreviousInput>());
             UserContextResolver.PreviousQuestions = questions;
         }
 
@@ -77,7 +77,7 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual.Actions
             }
         }
 
-        private async Task<PreviousQuestion> AbstractPreviousQuestionItemsAsync(ITurnContext turnContext)
+        private async Task<PreviousInput> AbstractPreviousQuestionItemsAsync(ITurnContext turnContext)
         {
             var skillStateAccessor = ConversationState.CreateProperty<dynamic>(string.Format("{0}State", SkillName));
             var skillState = await skillStateAccessor.GetAsync(turnContext);
@@ -85,7 +85,7 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual.Actions
             string intent = skillState.LuisResult.TopIntent().Item1.ToString();
             DateTimeOffset timeStamp = turnContext.Activity.Timestamp ?? new DateTimeOffset();
 
-            return new PreviousQuestion()
+            return new PreviousInput()
             {
                 Utterance = utterance,
                 Intent = intent,
@@ -93,7 +93,7 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual.Actions
             };
         }
 
-        private async Task ExecuteSavePreviousQuestionAsync(ITurnContext turnContext, PreviousQuestion newQuestion)
+        private async Task ExecuteSavePreviousQuestionAsync(ITurnContext turnContext, PreviousInput newQuestion)
         {
             if (newQuestion == null)
             {
@@ -106,8 +106,8 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual.Actions
                 return;
             }
 
-            var questionAccessor = UserState.CreateProperty<List<PreviousQuestion>>(string.Format("{0}Questions", SkillName));
-            var previousQuestions = await questionAccessor.GetAsync(turnContext, () => new List<PreviousQuestion>());
+            var questionAccessor = UserState.CreateProperty<List<PreviousInput>>(string.Format("{0}Questions", SkillName));
+            var previousQuestions = await questionAccessor.GetAsync(turnContext, () => new List<PreviousInput>());
 
             // If already exists, refresh timestamp.
             var duplicateQuestion = previousQuestions.Where(x => x.Utterance == newQuestion.Utterance).ToList();
@@ -129,18 +129,18 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual.Actions
             }
         }
 
-        private IReplacementStrategy<PreviousQuestion> GetStrategy(ReplacementStrategy replacementStrategy)
+        private IReplacementStrategy<PreviousInput> GetStrategy(ReplacementStrategy replacementStrategy)
         {
             switch (replacementStrategy)
             {
                 case ReplacementStrategy.FIFO:
-                    return new FIFOStrategy<PreviousQuestion>();
+                    return new FIFOStrategy<PreviousInput>();
                 case ReplacementStrategy.LRU:
-                    return new LRUStrategy<PreviousQuestion>();
+                    return new LRUStrategy<PreviousInput>();
                 case ReplacementStrategy.Random:
-                    return new RandomStrategy<PreviousQuestion>();
+                    return new RandomStrategy<PreviousInput>();
                 default:
-                    return new FIFOStrategy<PreviousQuestion>();
+                    return new FIFOStrategy<PreviousInput>();
             }
         }
