+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Models/PreviousInput.cs
@@ -2,7 +2,7 @@
 
 namespace Microsoft.Bot.Builder.Solutions.Contextual.Models
 {
-    public class PreviousQuestion
+    public class PreviousInput
     {
         public string Utterance { get; set; }
