+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Services/Sentiment/DocumentSentimentLabel.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Solutions.Contextual.Services.Sentiment
+{
+    public enum DocumentSentimentLabel
+    {
+        Positive,
+
+        Neutral,
+
+        Negative,
+
+        Mixed,
+    }
+}
