+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Services/Sentiment/SentimentConfidenceScoreLabel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Solutions.Contextual.Services.Sentiment
+{
+    public class SentimentConfidenceScoreLabel
+    {
+        public double Positive { get; set; }
+
+        public double Neutral { get; set; }
+
+        public double Negative { get; set; }
+    }
+}
