+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Services/Sentiment/TextAnalyticsBatchInput.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Solutions.Contextual.Services.Sentiment
+{
+    public class TextAnalyticsBatchInput
+    {
+        public IList<TextAnalyticsInput> Documents { get; set; }
+    }
+}
