+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Services/Sentiment/TextAnalyticsInput.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Solutions.Contextual.Services.Sentiment
+{
+    public class TextAnalyticsInput
+    {
+        /// <summary>
+        /// A unique, non-empty document identifier.
+        /// </summary>
+        public string Id { get; set; }
+
+        /// <summary>
+        /// The input text to process.
+        /// </summary>
+        public string Text { get; set; }
+
+        /// <summary>
+        /// The language code. Default is english ("en").
+        /// </summary>
+        public string LanguageCode { get; set; } = "en";
+    }
+}
