+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/UserContextManager.cs
@@ -19,7 +19,7 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual
 
         public static int DialogIndex { get; set; } = 0;
 
-        internal List<PreviousQuestion> PreviousQuestions { get; set; } = new List<PreviousQuestion>();
+        internal List<PreviousInput> PreviousQuestions { get; set; } = new List<PreviousInput>();
 
         public async Task<IList<string>> GetResolvedContactAsync(RelatedEntityInfo relatedEntityInfo)
         {
@@ -45,7 +45,7 @@ namespace Microsoft.Bot.Builder.Solutions.Contextual
             DialogIndex++;
         }
 
-        public List<PreviousQuestion> GetPreviousQuestions()
+        public List<PreviousInput> GetPreviousQuestions()
         {
             return PreviousQuestions;
         }
